pyproject.toml
[project]
name = "watsonx-autodeploy"
version = "0.1.0"
description = "Automated deployment library for IBM WatsonX AI services with LangGraph support"
readme = "README.md"
requires-python = ">=3.11"
authors = [
    {name = "Nicholas Renotte", email = "nick@nichollasrenotte.com"}
]
license = {text = "MIT"}
keywords = ["watsonx", "ai", "deployment", "langgraph", "automation"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "ibm-watsonx-ai>=1.3.34",
    "langchain>=0.3.27",
    "langchain-ibm>=0.3.15",
    "langgraph>=0.6.5",
    "python-dotenv>=1.1.1",
    "yfinance>=0.2.65",
]

[project.urls]
Homepage = "https://github.com/nicholasrenotte/watsonx-autodeploy"
Repository = "https://github.com/nicholasrenotte/watsonx-autodeploy.git"
Issues = "https://github.com/nicholasrenotte/watsonx-autodeploy/issues"

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "black>=23.0",
    "flake8>=6.0",
    "mypy>=1.0",
]

# pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-dir]
"" = "src"
